refactor: reorganize dashboard service imports and remove obsolete files
- Updated import paths for payroll, advance, expense, and timesheet aggregate functions to reflect new service structure.
- Deleted outdated aggregate functions and their corresponding files to streamline the codebase and improve maintainability.
- Enhanced clarity in AGENTS.md by adding a new section for dashboard overview chart aggregates.
- Cleaned up unused imports and ensured consistency across dashboard components.
