+++ AGENTS.md
@@ -117,6 +117,7 @@ Drizzle table definitions live in `db/tables/` and are re-exported from `db/sche
 | Route pages + server actions                     | `app/dashboard/<feature>/`                                                                                                                                                                                                                                                        |
 | API routes and shared transport helpers          | `app/api/`, `app/api/_shared/`                                                                                                                                                                                                                                                    |
 | Shared service/use-case modules                  | `services/<feature>/`                                                                                                                                                                                                                                                             |
+| Dashboard overview chart aggregates              | `services/dashboard/` (`timesheet-monthly-hours-aggregates.ts`, `expense-monthly-aggregates.ts`, `advance-monthly-repayment-aggregates.ts`, `payroll-monthly-category-aggregates.ts`, `payroll-monthly-grand-total-aggregates.ts`)                                                |
 | Worker create/update service                     | `services/worker/save-worker.ts`, `services/worker/worker-upsert-rows.ts`                                                                                                                                                                                                         |
 | React PDF voucher templates                      | `services/pdf/react-pdf/`                                                                                                                                                                                                                                                         |
 | Supabase Storage helpers (documents bucket)      | `lib/supabase/storage.ts`                                                                                                                                                                                                                                                         |
